Deduplicate implementations for PriceGetter trait
