+++ runtime/common/Cargo.toml
@@ -35,6 +35,7 @@ dia-oracle = { git = "https://github.com/pendulum-chain/oracle-pallet", default-
 zenlink-protocol = { git = "https://github.com/zenlinkpro/Zenlink-DEX-Module", default-features = false, branch = "polkadot-v0.9.42" }
 
 spacewalk-primitives = { git = "https://github.com/pendulum-chain/spacewalk", default-features = false, rev ="e7a672ba1e21c98a70df30a6ee458317951dd597"}
+treasury-buyout-extension = { path = "../../pallets/treasury-buyout-extension", default-features = false }
 
 [features]
 default = [
@@ -60,6 +61,8 @@ std = [
     "zenlink-protocol/std",
     "spacewalk-primitives/std",
     "cumulus-primitives-core/std",
+    "treasury-buyout-extension/std",
+
 ]
 
 runtime-benchmarks = [
@@ -67,4 +70,6 @@ runtime-benchmarks = [
     "sp-runtime/runtime-benchmarks",
     "frame-support/runtime-benchmarks",
     "frame-system/runtime-benchmarks",
+    "treasury-buyout-extension/runtime-benchmarks",
+
 ]
